`fn get_random_number`: Remove redundant `as` casts.

+++ src/filmgrain.rs
@@ -205,7 +205,7 @@ fn get_random_number(bits: c_int, state: &mut c_uint) -> c_int {
     let bit = ((r >> 0 ^ r >> 1 ^ r >> 3 ^ r >> 12) & 1) as c_uint;
     *state = (r >> 1) as c_uint | bit << 15;
 
-    (*state >> 16 - bits & (((1 as c_int) << bits) - 1) as c_uint) as c_int
+    (*state >> 16 - bits & (1 << bits) - 1) as c_int
 }
 
 #[inline]
